requirements.txt
streamlit==1.37.1
tensorFlow==2.11.1
joblib==1.4.2
scikit-learn==1.1.3
Pillow==10.4.0